+++ src/grandorgue/settings-gui/GOSettingsOrgans.cpp
@@ -5,7 +5,7 @@
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
 
-#include "GOSettingsOrgan.h"
+#include "GOSettingsOrgans.h"
 
 #include <wx/button.h>
 #include <wx/listctrl.h>
